Extract `NextPageIndicator` so it can be reused
